Revisit parsing in main.go. Solve #2.
Hit an issue where client "sent" 0 bytes,
which then caused an empty read, which caused an error.

Ugly fix for now is reading a byte at a time until we have 9,
writing each into a buffer.
This is unfortunate, though, because it means our
read syscalls are at least 9x.
There's a better way to do this, I just need to be more awake to
do it well.