Solve #5 up to 5badname.test
Currently dealing with an issue with Scanner.Scan().
Basically, using Scan() means I'm unable to distinguish
`<data>\n` from `<data>EOF`, which isn't a big deal
*except* when a client sends username bytes but then
disconnects before sending \n.

We need to be aware of that \n in order to determine
if it should be forwarded to the server.